backend/requirements.txt
fastapi
uvicorn
python-dotenv
google-auth
google-auth-oauthlib
pydantic
pydantic-core
google-api-python-client
pytest
httpx
openai
beautifulsoup4
psycopg2-binary
sqlalchemy
playwright